Change label to be a whole number due to issues with the linear gauge displaying floats.

+++ Rion/Rion/ViewModels/FeedViewModel.cs
@@ -28,7 +28,7 @@ namespace Rion.ViewModels
             get => _voltage;
             set
             {
-                if (Math.Abs(_voltage - value) < 0.01) return;
+                if (Math.Abs(_voltage - value) < (double) 0.01) return;
                 _voltage = value;
                 OnPropertyChanged();
             }
